Update (and use) deps, simplifying config
This pulls in my standard eslint configuration, replacing the custom
config that was there before.  We also push TS build info files into
a specific location so we can avoid having them in either source or
destination directories.  This makes it easier to copy the project for
post build testing, and also simplifies packaging.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>