Rework the types that build the extension
This lets us avoid pulling in the original async definitions, which
gives us much more clarity about what types are available for (say)
auto-complete.  As far as I can tell, we're correctly building our type
now.

We also infer generic arguments from the types passed in, which leads to
cleaner code.  It may require explicit typing of the extension object,
but I'm happy to have that type there rather than in the call to extend.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>